Add `session_vars` to `mysql_query`. (#729)
This commit adds a `session_vars` dict to the `mysql_query` plugin,
similar to that done in #489. While this could also be done by using a
list of queries, having a dictionary allows for a cleaner query,
reusability (via merge key), and a more consistent experience when using
different plugins (like `mysql_user`, which supports `session_vars`).